requirements.txt
flask>=3.0.0
rdkit-pypi
werkzeug>=3.0.0
